requirement.txt
absl-py==0.10.0
appdirs==1.4.4
astroid==2.3.3
astunparse==1.6.3
attrs==18.1.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.1
computer==1.0.0
ConfigArgParse==1.0
coursera-dl==0.11.5
cv==1.0.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.8
imageio==2.9.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.16.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyter-http-over-ws==0.0.8
kaggle==1.5.8
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
mysql-connector-python==8.0.23
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.5
notebook==6.0.2
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
pandas==1.1.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.4.2
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.15.5
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==225
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.3.1
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
slugify==0.0.1
soupsieve==1.9.5
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.6.0
torchsummary==1.5.1
torchvision==0.7.0
tornado==6.0.3
tqdm==4.48.2
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.24.3
virtualenv==20.4.6
vision==1.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==2.0.1
zipp==0.6.0
